refactor: [FE] 결과스케치북에 사용된 컴포넌트 이동
component 하위 폴더인 resultSketchbook로 이동했어요.
