+++ frontend/src/components/resultSketchbook/QuizResultContent.tsx
